Move non-view function from views.services to helpers.services
