+++ app/__init__.py
@@ -3,7 +3,6 @@ import re
 from datetime import timedelta
 from flask import Flask, request, redirect, session
 from flask_login import LoginManager
-from flask._compat import string_types
 from flask_wtf.csrf import CsrfProtect
 
 from dmutils import apiclient, init_app, flask_featureflags
@@ -68,27 +67,3 @@ def config_attrs(config):
     """Returns config attributes from a Config object"""
     p = re.compile('^[A-Z_]+$')
     return filter(lambda attr: bool(p.match(attr)), dir(config))
-
-
-def convert_to_boolean(value):
-    """Turn strings to bools if they look like them
-
-    Truthy things should be True
-    >>> for truthy in ['true', 'on', 'yes', '1']:
-    ...   assert convert_to_boolean(truthy) == True
-
-    Falsey things should be False
-    >>> for falsey in ['false', 'off', 'no', '0']:
-    ...   assert convert_to_boolean(falsey) == False
-
-    Other things should be unchanged
-    >>> for value in ['falsey', 'other', True, 0]:
-    ...   assert convert_to_boolean(value) == value
-    """
-    if isinstance(value, string_types):
-        if value.lower() in ['t', 'true', 'on', 'yes', '1']:
-            return True
-        elif value.lower() in ['f', 'false', 'off', 'no', '0']:
-            return False
-
-    return value
