+++ app/main/views/services.py
@@ -2,7 +2,11 @@ from flask_login import login_required, current_user
 from flask import render_template, request, redirect, url_for, abort
 
 from ...main import main, existing_service_content, new_service_content
-from ... import data_api_client, flask_featureflags, convert_to_boolean
+from ..helpers.services import (
+    get_formatted_section_data, get_section_questions, get_section_error_messages,
+    is_service_modifiable, is_service_associated_with_supplier,
+)
+from ... import data_api_client, flask_featureflags
 from dmutils.apiclient import APIError, HTTPError
 from dmutils.presenters import Presenters
 
@@ -35,7 +39,7 @@ def list_services():
 def edit_service(service_id):
     service = data_api_client.get_service(service_id).get('services')
 
-    if not _is_service_associated_with_supplier(service):
+    if not is_service_associated_with_supplier(service):
         abort(404)
 
     return _update_service_status(service)
@@ -47,10 +51,10 @@ def edit_service(service_id):
 def update_service_status(service_id):
     service = data_api_client.get_service(service_id).get('services')
 
-    if not _is_service_associated_with_supplier(service):
+    if not is_service_associated_with_supplier(service):
         abort(404)
 
-    if not _is_service_modifiable(service):
+    if not is_service_modifiable(service):
         return _update_service_status(
             service,
             "Sorry, but this service isn't modifiable."
@@ -103,7 +107,7 @@ def edit_section(service_id, section_id):
         abort(404)
     service = service['services']
 
-    if not _is_service_associated_with_supplier(service):
+    if not is_service_associated_with_supplier(service):
         abort(404)
 
     content = existing_service_content.get_builder().filter(service)
@@ -120,10 +124,6 @@ def edit_section(service_id, section_id):
     )
 
 
-def get_section_questions(section):
-    return [question['id'] for question in section['questions']]
-
-
 @main.route('/services/<string:service_id>/edit/<string:section_id>', methods=['POST'])
 @login_required
 @flask_featureflags.is_active_feature('EDIT_SERVICE_PAGE')
@@ -133,7 +133,7 @@ def update_section(service_id, section_id):
         abort(404)
     service = service['services']
 
-    if not _is_service_associated_with_supplier(service):
+    if not is_service_associated_with_supplier(service):
         abort(404)
 
     content = existing_service_content.get_builder().filter(service)
@@ -141,7 +141,7 @@ def update_section(service_id, section_id):
     if section is None:
         abort(404)
 
-    posted_data = _get_formatted_section_data(section)
+    posted_data = get_formatted_section_data(section)
 
     try:
         data_api_client.update_service(
@@ -150,7 +150,7 @@ def update_section(service_id, section_id):
             current_user.email_address,
             "supplier app")
     except HTTPError as e:
-        errors_map = _get_section_error_messages(e, service['lot'])
+        errors_map = get_section_error_messages(e, service['lot'])
         if not posted_data.get('serviceName', None):
             posted_data['serviceName'] = service.get('serviceName', '')
         return render_template(
@@ -258,7 +258,7 @@ def create_new_draft_service():
 def copy_draft_service(service_id):
     draft = data_api_client.get_draft_service(service_id).get('services')
 
-    if not _is_service_associated_with_supplier(draft):
+    if not is_service_associated_with_supplier(draft):
         abort(404)
 
     try:
@@ -282,7 +282,7 @@ def view_service_submission(service_id):
     except HTTPError as e:
         abort(e.status_code)
 
-    if not _is_service_associated_with_supplier(draft):
+    if not is_service_associated_with_supplier(draft):
         abort(404)
 
     content = new_service_content.get_builder().filter(draft)
@@ -304,7 +304,7 @@ def edit_service_submission(service_id, section_id):
     except HTTPError as e:
         abort(e.status_code)
 
-    if not _is_service_associated_with_supplier(draft):
+    if not is_service_associated_with_supplier(draft):
         abort(404)
 
     content = new_service_content.get_builder().filter(draft)
@@ -330,7 +330,7 @@ def update_section_submission(service_id, section_id):
     except HTTPError as e:
         abort(e.status_code)
 
-    if not _is_service_associated_with_supplier(draft):
+    if not is_service_associated_with_supplier(draft):
         abort(404)
 
     content = new_service_content.get_builder().filter(draft)
@@ -338,7 +338,7 @@ def update_section_submission(service_id, section_id):
     if section is None:
         abort(404)
 
-    posted_data = _get_formatted_section_data(section)
+    posted_data = get_formatted_section_data(section)
     posted_data['page_questions'] = get_section_questions(section)
 
     try:
@@ -347,7 +347,7 @@ def update_section_submission(service_id, section_id):
             posted_data,
             current_user.email_address)
     except HTTPError as e:
-        errors_map = _get_section_error_messages(e, draft['lot'])
+        errors_map = get_section_error_messages(e, draft['lot'])
         if not posted_data.get('serviceName', None):
             posted_data['serviceName'] = draft.get('serviceName', '')
         return render_template(
@@ -366,59 +366,6 @@ def update_section_submission(service_id, section_id):
         return redirect(url_for(".edit_service_submission", service_id=service_id, section_id=next_section))
     else:
         return redirect(url_for(".view_service_submission", service_id=service_id))
-
-
-def _is_service_associated_with_supplier(service):
-    return service.get('supplierId') == current_user.supplier_id
-
-
-def _is_service_modifiable(service):
-    return service.get('status') != 'disabled'
-
-
-def _get_error_message(error, message_key, content):
-    validations = [
-        validation for validation in content.get_question(error)['validations']
-        if validation['name'] == message_key]
-
-    if len(validations):
-        return validations[0]['message']
-    else:
-        return 'There was a problem with the answer to this question'
-
-
-def _is_list_type(key):
-    """Return True if a given key is a list type"""
-    if key == 'serviceTypes':
-        return True
-    return new_service_content.get_question(key)['type'] in ['list', 'checkbox', 'pricing']
-
-
-def _is_boolean_type(key):
-    """Return True if a given key is a boolean type"""
-    return new_service_content.get_question(key)['type'] == 'boolean'
-
-
-def _get_formatted_section_data(section):
-    section_data = dict(list(request.form.items()))
-    section_data = _filter_keys(section_data, get_section_questions(section))
-    # Turn responses which have multiple parts into lists and booleans into booleans
-    for key in section_data:
-        if _is_list_type(key):
-            section_data[key] = request.form.getlist(key)
-        elif _is_boolean_type(key):
-            section_data[key] = convert_to_boolean(section_data[key])
-    return section_data
-
-
-def _filter_keys(data, keys):
-    """Return a dictionary filtered by a list of keys
-
-    >>> _filter_keys({'a': 1, 'b': 2}, ['a'])
-    {'a': 1}
-    """
-    key_set = set(keys) & set(data.keys())
-    return {key: data[key] for key in key_set}
 
 
 def _update_service_status(service, error_message=None):
@@ -453,24 +400,3 @@ def _update_service_status(service, error_message=None):
         sections=content,
         error=error_message,
         **dict(question, **template_data)), status_code
-
-
-def _get_section_error_messages(e, lot):
-    errors_map = {}
-    for error in e.message.keys():
-        if error == '_form':
-            abort(400, "Submitted data was not in a valid format")
-        else:
-            message_key = e.message[error]
-            if error == 'serviceTypes':
-                error = 'serviceTypes{}'.format(lot)
-            validation_message = _get_error_message(error,
-                                                    message_key,
-                                                    new_service_content)
-            question_id = new_service_content.get_question(error)['id']
-            errors_map[question_id] = {
-                'input_name': error,
-                'question': new_service_content.get_question(error)['question'],
-                'message': validation_message
-            }
-    return errors_map
